+++ src/server/RLCreatorLayer.hpp
@@ -15,6 +15,7 @@ class RLCreatorLayer : public CCLayer {
       void onNewRated(CCObject* sender);
       void onSendLayouts(CCObject* sender);
       void onInfoButton(CCObject* sender);
+      void onUnknownButton(CCObject* sender);
 
       CCNode* m_bgContainer = nullptr;
       std::vector<CCSprite*> m_bgTiles;